updated gh pages gen script

+++ create_gh_pages.sh
@@ -11,16 +11,21 @@
 # Requires the following var is set in the sphinx makefile:
 # GH_PAGES_SOURCES = source Makefile
 
-GH_PAGES_SOURCES="source Makefile"
-git checkout gh-pages
-#rm -rf build _sources _static
-touch ./.nojekyll
-git checkout master doc/source doc/Makefile
-git reset HEAD
-cd doc
-make html
-mv -fv build/html/* ../.
-rm -rf source Makefile build
-cd ..
-git add -A
-git commit -m "Gen gh-pages for `git log master -1 --pretty=short --abbrev-commit`" && git push origin gh-pages ; git checkout master
+if [-n "$git status --porcelain)" ]; then
+  echo "Master is clean. Building docs...";
+  GH_PAGES_SOURCES="source Makefile"
+  git checkout gh-pages
+  rm -rf build _sources _static
+  touch ./.nojekyll
+  git checkout master doc/source doc/Makefile
+  git reset HEAD
+  cd doc
+  make html
+  mv -fv build/html/* ../.
+  rm -rf source Makefile build
+  cd ..
+  git add -A
+  git commit -m "Gen gh-pages for `git log master -1 --pretty=short --abbrev-commit`" && git push origin gh-pages ; git checkout master
+else
+  echo "Commit master changes before building docs.";
+fi
